code cleanup changes and others due to v20.11->v21.0 upgrade changes
