--- systest/ldbc/upgrade_helper_test.go
@@ -1,48 +1,0 @@
-//go:build upgrade
-
-/*
- * Copyright 2023 Dgraph Labs, Inc. and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package main
-
-import (
-	"os"
-	"path/filepath"
-
-	"github.com/dgraph-io/dgraph/dgraphtest"
-)
-
-func (lsuite *LdbcTestSuite) bulkLoader() error {
-	dataDir := os.Getenv("TEST_DATA_DIRECTORY")
-	rdfFile := dataDir
-	schemaFile := filepath.Join(dataDir, "ldbcTypes.schema")
-	return lsuite.lc.BulkLoad(dgraphtest.BulkOpts{
-		DataFiles:   []string{rdfFile},
-		SchemaFiles: []string{schemaFile},
-	})
-}
-
-func (lsuite *LdbcTestSuite) StartAlpha() error {
-	c := lsuite.lc
-	if err := c.Start(); err != nil {
-		return err
-	}
-	return c.HealthCheck(false)
-}
-
-func (lsuite *LdbcTestSuite) StopAlphasForCoverage() {
-	lsuite.lc.StopAlpha(0)
-}
